Update README.md
What I like about the film is it talk a lot of struggles of being a  students also this movie is for everyone and it has a lot of message for everybody. I like on how it deals with suicides, in todays generation there are a lot of students who commits suicide but the message of this film is not about taking your own life, no matter how heavy the burdens you carry in life.